add docs to builder
