+++ Cargo.toml
@@ -61,7 +61,6 @@ uuid = { version = "0.8.1", features = ["serde", "v4"] }
 serde_bytes = "0.11.5"
 serde_with = { version = "1", optional = true }
 time = { version = "0.3.9", features = ["formatting", "parsing", "macros", "large-dates"] }
-typed-builder = "0.10.0"
 
 [dev-dependencies]
 assert_matches = "1.2"
